window_identifier: Do not destroy exported surface
The other side of the bus should take ownership of the exported toplevel
and destroy it, otherwise things might go badly. This is something that
should be called only once the associated GtkWindow (for example) is
destroyed but we don't have a nice api for this.

One can check that destroying the exported toplevel too quick causes the file chooser to
not be modal, but due to sheer luck it works with the current
implementation of the file chooser, other portals (or compositors) might not be as lucky.

We also use Option<String> to transport the handle, this ensures we don't
have an empty string comming from Default::default().