documents: Take `AppId` or `OptionAppId` instead of `TryInto` generic
Some of these functions accept an empty string, but calling with that
was erroring since those are not valid as app ids. So accepting an
`Option` would be idiomatic in Rust.

It could take `Option<T>` where `T` has a `TryFrom` bound, but that's an
awkward API to deal with, since `None` can't be passed without a type.
It also seems unidiomatic to take a generic argument when the conversion
is failable; it would be cleaner to leave failable conversion to the
caller.

Fixes https://github.com/bilelmoussaoui/ashpd/issues/129.