Drop async-std crate
The async-std crate didn't had a release for 1 1/2 years. It also pulls in
a lot of dependencies that are usually not needed. The Path::exists()
feature of async-std is just using spawn_blocking around the std
metadata function and therefore settling on a specific executor and
spawning a bunch of threads that might not be needed.

Therefore I replaced all uses of async-std with the crates from which
async-std is build and that are well maintained.