file-chooser: Allow to add filters
Add a combox box that allows select different filters to the OpenFile
and SaveFile dialogs. Adding different mime types allows for some
flexibility when testing e.g. file choser implementations.

When a filter set of filters is selected we pick the first filter as the
`current_filter`.

The default is no filters.

Signed-off-by: Guido Günther <agx@sigxcpu.org>