backend: Pass handle token to method calls
By having a handle token, an implementer can now associate further calls
on the `Request` with its backend logic.

For example, suppose an app A calls `Account.GetUserInformation()`
followed by an app B. The implementer shows a dialog for each of them
that lets the user give their response. Now, if any of these apps call
`Request.close()`, then the implementer does not know which dialog to
close, as it did not have any way to associate the `Request` to its
backend logic in the first place.

This commit solves this issue.