wip: client: Add zbus_macros from workspace
This is a dependency of zbus anyways.

Otherwise compilation fails with:

```
error[E0004]: non-exhaustive patterns: `&zvariant_utils::signature::Signature::Maybe(_)` not covered
   --> /var/home/deathwish/.local/share/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/zvariant-5.8.0/src/de.rs:198:11
    |
198 |     match signature {
    |           ^^^^^^^^^ pattern `&zvariant_utils::signature::Signature::Maybe(_)` not covered
    |
note: `zvariant_utils::signature::Signature` defined here
   --> /var/home/deathwish/.local/share/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/zvariant_utils-3.2.1/src/signature/mod.rs:42:1
    |
 42 | pub enum Signature {
    | ^^^^^^^^^^^^^^^^^^
...
 99 |     Maybe(Child),
    |     ----- not covered
    = note: the matched value is of type `&zvariant_utils::signature::Signature`
help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
217 ~         Signature::Structure { .. } => de.deserialize_seq(visitor),
218 ~         &zvariant_utils::signature::Signature::Maybe(_) => todo!(),
    |
```

wip: Add zbus_macros to client instead

This also fixes the issue and does not add a dependency that was not
there from the start.